backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
sentence-transformers
scikit-learn
reportlab
pandas
google-genai